[test/#315-slack] v3으로 테스트 코드 수정
