+++ slack-service/src/test/java/com/business/slackservice/slack/presentation/controller/SlackControllerApiV3Test.java
@@ -6,8 +6,9 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.pr
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
 
-import com.business.slackservice.application.dto.v1.request.slackTemplate.ReqSlackTemplatePostDTOApiV1;
-import com.business.slackservice.application.dto.v1.request.slackTemplate.ReqSlackTemplatePutDTOApiV1;
+import com.business.slackservice.application.dto.v3.request.slack.ReqSlackPostDTOApiV3;
+import com.business.slackservice.application.dto.v3.request.slack.ReqSlackPostDTOApiV3.Slack;
+import com.business.slackservice.domain.slack.vo.TargetType;
 import com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper;
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.epages.restdocs.apispec.SimpleType;
@@ -30,7 +31,7 @@ import org.springframework.transaction.annotation.Transactional;
 @AutoConfigureMockMvc
 @Transactional
 @ActiveProfiles("dev")
-public class SlackTemplateControllerApiV1Test {
+public class SlackControllerApiV3Test {
 
     @Autowired
     private MockMvc mockMvc;
@@ -38,18 +39,26 @@ public class SlackTemplateControllerApiV1Test {
     @Autowired
     private ObjectMapper objectMapper;
 
+    private final UUID slackId = UUID.fromString("1f8f8103-d3bd-4b88-b51b-5e40012b6263");
+
     @Test
-    public void testSlackTemplatePostSuccess() throws Exception {
-        ReqSlackTemplatePostDTOApiV1 req = ReqSlackTemplatePostDTOApiV1.builder()
-            .slackTemplate(
-                ReqSlackTemplatePostDTOApiV1.SlackTemplate.builder()
-                    .slackEventTypeId(UUID.randomUUID())
-                    .content("{{relatedName}} 주문이 정상적으로 접수되었습니다.}")
+    public void testSlackPostSuccess() throws Exception {
+        ReqSlackPostDTOApiV3 req = ReqSlackPostDTOApiV3.builder()
+            .slack(
+                ReqSlackPostDTOApiV3.Slack.builder()
+                    .slackEventType("ORDER_COMPLETE")
+                    .relatedName("Slack-Related-Name")
+                    .target(
+                        Slack.SlackTarget.builder()
+                            .slackId("Slack-Id")
+                            .type(TargetType.USER_DM)
+                            .build()
+                    )
                     .build()
             )
             .build();
         mockMvc.perform(
-                RestDocumentationRequestBuilders.post("/v1/slack-templates")
+                RestDocumentationRequestBuilders.post("/v1/internal/slacks")
                     .content(dtoToJson(req))
                     .contentType(MediaType.APPLICATION_JSON)
             )
@@ -57,11 +66,11 @@ public class SlackTemplateControllerApiV1Test {
                 MockMvcResultMatchers.status().isOk()
             )
             .andDo(
-                MockMvcRestDocumentationWrapper.document("슬랙 템플릿 생성 성공",
+                MockMvcRestDocumentationWrapper.document("슬랙 생성 성공",
                     preprocessRequest(prettyPrint()),
                     preprocessResponse(prettyPrint()),
                     resource(ResourceSnippetParameters.builder()
-                        .tag("SLACK TEMPLATE v1")
+                        .tag("SLACK v3")
                         .build()
                     )
                 )
@@ -69,22 +78,22 @@ public class SlackTemplateControllerApiV1Test {
     }
 
     @Test
-    public void testSlackTemplateGetByIdSuccess() throws Exception {
+    public void testSlackGetByIdSuccess() throws Exception {
         mockMvc.perform(
-                RestDocumentationRequestBuilders.get("/v1/slack-templates/{id}", UUID.randomUUID())
+                RestDocumentationRequestBuilders.get("/v1/slacks/{id}", slackId)
             )
             .andExpectAll(
                 MockMvcResultMatchers.status().isOk()
             )
             .andDo(
-                MockMvcRestDocumentationWrapper.document("슬랙 템플릿 단건 조회 성공",
+                MockMvcRestDocumentationWrapper.document("슬랙 단건 조회 성공",
                     preprocessRequest(prettyPrint()),
                     preprocessResponse(prettyPrint()),
                     resource(ResourceSnippetParameters.builder()
-                        .tag("SLACK TEMPLATE v1")
+                        .tag("SLACK v3")
                         .pathParameters(
                             parameterWithName("id").type(SimpleType.STRING)
-                                .description("슬랙 템플릿 ID")
+                                .description("슬랙 ID")
                         )
                         .build()
                     )
@@ -93,19 +102,19 @@ public class SlackTemplateControllerApiV1Test {
     }
 
     @Test
-    public void testSlackTemplateGetSuccess() throws Exception {
+    public void testSlackGetSuccess() throws Exception {
         mockMvc.perform(
-                RestDocumentationRequestBuilders.get("/v1/slack-templates")
+                RestDocumentationRequestBuilders.get("/v1/slacks")
             )
             .andExpectAll(
                 MockMvcResultMatchers.status().isOk()
             )
             .andDo(
-                MockMvcRestDocumentationWrapper.document("슬랙 템플릿 목록 조회 성공",
+                MockMvcRestDocumentationWrapper.document("슬랙 목록 조회 성공",
                     preprocessRequest(prettyPrint()),
                     preprocessResponse(prettyPrint()),
                     resource(ResourceSnippetParameters.builder()
-                        .tag("SLACK TEMPLATE v1")
+                        .tag("SLACK v3")
                         .build()
                     )
                 )
@@ -113,58 +122,23 @@ public class SlackTemplateControllerApiV1Test {
     }
 
     @Test
-    public void testSlackTemplatePutByIdSuccess() throws Exception {
-        ReqSlackTemplatePutDTOApiV1 req = ReqSlackTemplatePutDTOApiV1.builder()
-            .slackTemplate(
-                ReqSlackTemplatePutDTOApiV1.SlackTemplate.builder()
-                    .SlackEventTypeId(UUID.randomUUID())
-                    .content("{{relatedName}} 주문이 정상적으로 접수되었습니다. 업데이트")
-                    .build()
-            )
-            .build();
-
+    public void testSlackDeleteByIdSuccess() throws Exception {
         mockMvc.perform(
-                RestDocumentationRequestBuilders.put("/v1/slack-templates/{id}", UUID.randomUUID())
-                    .content(dtoToJson(req))
-                    .contentType(MediaType.APPLICATION_JSON)
-            )
-            .andExpectAll(
-                MockMvcResultMatchers.status().isOk()
-            )
-            .andDo(
-                MockMvcRestDocumentationWrapper.document("슬랙 템플릿 수정 성공",
-                    preprocessRequest(prettyPrint()),
-                    preprocessResponse(prettyPrint()),
-                    resource(ResourceSnippetParameters.builder()
-                        .tag("SLACK TEMPLATE v1")
-                        .pathParameters(
-                            parameterWithName("id").type(SimpleType.STRING)
-                                .description("슬랙 템플릿 ID")
-                        )
-                        .build()
-                    )
-                )
-            );
-    }
-
-    @Test
-    public void testSlackTemplateDeleteByIdSuccess() throws Exception {
-        mockMvc.perform(
-                RestDocumentationRequestBuilders.delete("/v1/slack-templates/{id}", UUID.randomUUID())
+                RestDocumentationRequestBuilders.delete("/v1/slacks/{id}", slackId)
                     .header("X-User-Id", 1L)
             )
             .andExpectAll(
                 MockMvcResultMatchers.status().isOk()
             )
             .andDo(
-                MockMvcRestDocumentationWrapper.document("슬랙 템플릿 삭제 성공",
+                MockMvcRestDocumentationWrapper.document("슬랙 삭제 성공",
                     preprocessRequest(prettyPrint()),
                     preprocessResponse(prettyPrint()),
                     resource(ResourceSnippetParameters.builder()
-                        .tag("SLACK TEMPLATE v1")
+                        .tag("SLACK v3")
                         .pathParameters(
                             parameterWithName("id").type(SimpleType.STRING)
-                                .description("슬랙 템플릿 ID")
+                                .description("슬랙 ID")
                         )
                         .build()
                     )
