+++ slack-service/src/test/java/com/business/slackservice/slack/presentation/controller/SlackTemplateControllerApiV3Test.java
@@ -6,9 +6,8 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.pr
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
 
-import com.business.slackservice.application.dto.v1.request.slack.ReqSlackPostDTOApiV1;
-import com.business.slackservice.application.dto.v1.request.slack.ReqSlackPostDTOApiV1.Slack;
-import com.business.slackservice.domain.slack.vo.TargetType;
+import com.business.slackservice.application.dto.v3.request.slackTemplate.ReqSlackTemplatePostDTOApiV3;
+import com.business.slackservice.application.dto.v3.request.slackTemplate.ReqSlackTemplatePutDTOApiV3;
 import com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper;
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.epages.restdocs.apispec.SimpleType;
@@ -31,7 +30,7 @@ import org.springframework.transaction.annotation.Transactional;
 @AutoConfigureMockMvc
 @Transactional
 @ActiveProfiles("dev")
-public class SlackControllerApiV1Test {
+public class SlackTemplateControllerApiV3Test {
 
     @Autowired
     private MockMvc mockMvc;
@@ -39,25 +38,21 @@ public class SlackControllerApiV1Test {
     @Autowired
     private ObjectMapper objectMapper;
 
+    private final UUID slackEventTypeId = UUID.fromString("3abeba8a-8eba-4819-90a4-5ac21cc27382");
+    private final UUID slackTemplateId = UUID.fromString("1437ce66-9bfe-4a91-ac92-ea2a993e86a2");
+
     @Test
-    public void testSlackPostSuccess() throws Exception {
-        ReqSlackPostDTOApiV1 req = ReqSlackPostDTOApiV1.builder()
-            .slack(
-                ReqSlackPostDTOApiV1.Slack.builder()
-                    .slackEventType("Slack-Event-Type")
-                    .relatedName("Slack-Related-Name")
-                    .slackEventType("USER_DM")
-                    .target(
-                        Slack.SlackTarget.builder()
-                            .slackId("Slack-Id")
-                            .type(TargetType.USER_DM)
-                            .build()
-                    )
+    public void testSlackTemplatePostSuccess() throws Exception {
+        ReqSlackTemplatePostDTOApiV3 req = ReqSlackTemplatePostDTOApiV3.builder()
+            .slackTemplate(
+                ReqSlackTemplatePostDTOApiV3.SlackTemplate.builder()
+                    .slackEventTypeId(slackEventTypeId)
+                    .content("{{relatedName}} 주문이 정상적으로 접수되었습니다.}")
                     .build()
             )
             .build();
         mockMvc.perform(
-                RestDocumentationRequestBuilders.post("/v1/internal/slacks")
+                RestDocumentationRequestBuilders.post("/v1/slack-templates")
                     .content(dtoToJson(req))
                     .contentType(MediaType.APPLICATION_JSON)
             )
@@ -65,11 +60,11 @@ public class SlackControllerApiV1Test {
                 MockMvcResultMatchers.status().isOk()
             )
             .andDo(
-                MockMvcRestDocumentationWrapper.document("슬랙 생성 성공",
+                MockMvcRestDocumentationWrapper.document("슬랙 템플릿 생성 성공",
                     preprocessRequest(prettyPrint()),
                     preprocessResponse(prettyPrint()),
                     resource(ResourceSnippetParameters.builder()
-                        .tag("SLACK v1")
+                        .tag("SLACK TEMPLATE v3")
                         .build()
                     )
                 )
@@ -77,22 +72,22 @@ public class SlackControllerApiV1Test {
     }
 
     @Test
-    public void testSlackGetByIdSuccess() throws Exception {
+    public void testSlackTemplateGetByIdSuccess() throws Exception {
         mockMvc.perform(
-                RestDocumentationRequestBuilders.get("/v1/slacks/{id}", UUID.randomUUID())
+                RestDocumentationRequestBuilders.get("/v1/slack-templates/{id}", slackTemplateId)
             )
             .andExpectAll(
                 MockMvcResultMatchers.status().isOk()
             )
             .andDo(
-                MockMvcRestDocumentationWrapper.document("슬랙 단건 조회 성공",
+                MockMvcRestDocumentationWrapper.document("슬랙 템플릿 단건 조회 성공",
                     preprocessRequest(prettyPrint()),
                     preprocessResponse(prettyPrint()),
                     resource(ResourceSnippetParameters.builder()
-                        .tag("SLACK v1")
+                        .tag("SLACK TEMPLATE v3")
                         .pathParameters(
                             parameterWithName("id").type(SimpleType.STRING)
-                                .description("슬랙 ID")
+                                .description("슬랙 템플릿 ID")
                         )
                         .build()
                     )
@@ -101,22 +96,53 @@ public class SlackControllerApiV1Test {
     }
 
     @Test
-    public void testSlackGetSuccess() throws Exception {
+    public void testSlackTemplateGetSuccess() throws Exception {
         mockMvc.perform(
-                RestDocumentationRequestBuilders.get("/v1/slacks/{id}", UUID.randomUUID())
+                RestDocumentationRequestBuilders.get("/v1/slack-templates")
             )
             .andExpectAll(
                 MockMvcResultMatchers.status().isOk()
             )
             .andDo(
-                MockMvcRestDocumentationWrapper.document("슬랙 삭제 성공",
+                MockMvcRestDocumentationWrapper.document("슬랙 템플릿 목록 조회 성공",
                     preprocessRequest(prettyPrint()),
                     preprocessResponse(prettyPrint()),
                     resource(ResourceSnippetParameters.builder()
-                        .tag("SLACK v1")
+                        .tag("SLACK TEMPLATE v3")
+                        .build()
+                    )
+                )
+            );
+    }
+
+    @Test
+    public void testSlackTemplatePutByIdSuccess() throws Exception {
+        ReqSlackTemplatePutDTOApiV3 req = ReqSlackTemplatePutDTOApiV3.builder()
+            .slackTemplate(
+                ReqSlackTemplatePutDTOApiV3.SlackTemplate.builder()
+                    .SlackEventTypeId(slackEventTypeId)
+                    .content("{{relatedName}} 주문이 정상적으로 접수되었습니다. 업데이트")
+                    .build()
+            )
+            .build();
+
+        mockMvc.perform(
+                RestDocumentationRequestBuilders.put("/v1/slack-templates/{id}", slackTemplateId)
+                    .content(dtoToJson(req))
+                    .contentType(MediaType.APPLICATION_JSON)
+            )
+            .andExpectAll(
+                MockMvcResultMatchers.status().isOk()
+            )
+            .andDo(
+                MockMvcRestDocumentationWrapper.document("슬랙 템플릿 수정 성공",
+                    preprocessRequest(prettyPrint()),
+                    preprocessResponse(prettyPrint()),
+                    resource(ResourceSnippetParameters.builder()
+                        .tag("SLACK TEMPLATE v3")
                         .pathParameters(
                             parameterWithName("id").type(SimpleType.STRING)
-                                .description("슬랙 ID")
+                                .description("슬랙 템플릿 ID")
                         )
                         .build()
                     )
@@ -125,9 +151,9 @@ public class SlackControllerApiV1Test {
     }
 
     @Test
-    public void testSlackDeleteByIdSuccess() throws Exception {
+    public void testSlackTemplateDeleteByIdSuccess() throws Exception {
         mockMvc.perform(
-                RestDocumentationRequestBuilders.delete("/v1/slacks/{id}", UUID.randomUUID())
+                RestDocumentationRequestBuilders.delete("/v1/slack-templates/{id}", slackTemplateId)
                     .header("X-User-Id", 1L)
             )
             .andExpectAll(
@@ -138,10 +164,10 @@ public class SlackControllerApiV1Test {
                     preprocessRequest(prettyPrint()),
                     preprocessResponse(prettyPrint()),
                     resource(ResourceSnippetParameters.builder()
-                        .tag("SLACK v1")
+                        .tag("SLACK TEMPLATE v3")
                         .pathParameters(
                             parameterWithName("id").type(SimpleType.STRING)
-                                .description("슬랙 ID")
+                                .description("슬랙 템플릿 ID")
                         )
                         .build()
                     )
